Thay code

+++ be/src/main/java/training/g2/repository/StockInItemRepository.java
@@ -1,7 +1,10 @@
 package training.g2.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import training.g2.model.StockIn;
 import training.g2.model.StockInItem;
 
-public interface StockInItemRepository extends JpaRepository<StockInItem, Integer> {
+import java.util.List;
+
+public interface StockInItemRepository extends JpaRepository<StockInItem, Long> {
 }
